加入管理页面 修复全局异常处理和接口统一返回

+++ blog-api/src/main/java/com/erotsx/blog/vo/ResultCode.java
@@ -7,7 +7,7 @@ public enum ResultCode implements ErrorCode {
     UNAUTHORIZED(401, "暂未登录或token已经过期"),
 
     FORBIDDEN(403, "没有相关权限");
-    private int code;
+    private long code;
     private String message;
 
     ResultCode(int code, String message) {
